+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## next
 
+*Bug fixes*
+- Attempting to deploy from a directory that only contains `secrets.ejson` will no longer fail deploy ([#416](https://github.com/Shopify/kubernetes-deploy/pull/416))
+
 ## 0.24.0
 
 *Features*